Extract common result APIs to Applicable module
